+++ examples/hot-iced/lib/src/lib.rs
@@ -1,4 +1,4 @@
-use iced::{widget::Text, Command, Element};
+use iced::{Command, Element, widget::Text};
 use std::time::Instant;
 
 #[derive(Debug, Clone, Copy)]
@@ -24,7 +24,7 @@ impl Default for State {
     }
 }
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub fn update(state: &mut State, message: Message) -> Command<Message> {
     match message {
         Message::Tick(instant) => {
@@ -35,7 +35,7 @@ pub fn update(state: &mut State, message: Message) -> Command<Message> {
     Command::none()
 }
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub fn view<'a>(state: &State) -> Element<'a, Message> {
     Text::new(format!("The time is {:?}!", state.time)).into()
 }