+++ macro-no-mangle-if-debug/src/lib.rs
@@ -1,4 +1,4 @@
-//! Will add `#[no_mangle]` to the item it is applied but only in debug mode.
+//! Will add `#[unsafe(no_mangle)]` to the item it is applied but only in debug mode.
 //!
 //! This is useful for use with [hot-lib-reloader](https://crates.io/crates/hot-lib-reloader) to conditionally expose library functions to the lib reloader only in debug mode.
 //! In release mode where a build is to be expected fully static, no additional penalty is paid.
@@ -12,7 +12,7 @@
 //!
 //! ```xxx
 //! #[cfg(debug_assertions)]
-//! #[no_mangle]
+//! #[unsafe(no_mangle)]
 //! fn func() {}
 //!
 //! #[cfg(not(debug_assertions))]
@@ -20,8 +20,8 @@
 //! ```
 
 use proc_macro2::TokenStream;
-use quote::{quote, ToTokens};
-use syn::{parse_macro_input, token};
+use quote::{ToTokens, quote};
+use syn::{parse_macro_input, punctuated::Punctuated, token};
 
 /// See package doc.
 #[proc_macro_attribute]
@@ -34,15 +34,13 @@ pub fn no_mangle_if_debug(
 
     debug_item.attrs.push(create_attribute(
         "cfg",
-        quote! { (debug_assertions) }.into_token_stream(),
+        quote! { debug_assertions }.into_token_stream(),
     ));
-    debug_item
-        .attrs
-        .push(create_attribute("no_mangle", Default::default()));
+    debug_item.attrs.push(create_unsafe_attribute("no_mangle"));
 
     release_item.attrs.push(create_attribute(
         "cfg",
-        quote! { (not(debug_assertions)) }.into_token_stream(),
+        quote! { not(debug_assertions) }.into_token_stream(),
     ));
 
     (quote! {
@@ -52,21 +50,46 @@ pub fn no_mangle_if_debug(
     .into()
 }
 
-fn create_attribute(ident: &str, tokens: TokenStream) -> syn::Attribute {
+fn create_unsafe_attribute(ident: &str) -> syn::Attribute {
     let span = proc_macro2::Span::call_site();
+    let mut segments = Punctuated::new();
+    segments.push(syn::PathSegment {
+        ident: syn::Ident::new("unsafe", span),
+        arguments: syn::PathArguments::None,
+    });
     syn::Attribute {
         style: syn::AttrStyle::Outer,
         pound_token: token::Pound { spans: [span] },
         bracket_token: token::Bracket::default(),
-        path: syn::Path {
-            leading_colon: None,
-            segments: [syn::PathSegment {
-                ident: syn::Ident::new(ident, span),
-                arguments: syn::PathArguments::None,
-            }]
-            .into_iter()
-            .collect(),
-        },
-        tokens,
+        meta: syn::Meta::List(syn::MetaList {
+            path: syn::Path {
+                leading_colon: None,
+                segments,
+            },
+            delimiter: syn::MacroDelimiter::Paren(syn::token::Paren::default()),
+            tokens: syn::Ident::new(ident, span).into_token_stream(),
+        }),
+    }
+}
+
+fn create_attribute(ident: &str, tokens: TokenStream) -> syn::Attribute {
+    let span = proc_macro2::Span::call_site();
+    let mut segments = Punctuated::new();
+    segments.push(syn::PathSegment {
+        ident: syn::Ident::new(ident, span),
+        arguments: syn::PathArguments::None,
+    });
+    syn::Attribute {
+        style: syn::AttrStyle::Outer,
+        pound_token: token::Pound { spans: [span] },
+        bracket_token: token::Bracket::default(),
+        meta: syn::Meta::List(syn::MetaList {
+            path: syn::Path {
+                leading_colon: None,
+                segments,
+            },
+            delimiter: syn::MacroDelimiter::Paren(syn::token::Paren::default()),
+            tokens,
+        }),
     }
 }