+++ macro/examples/syn-test.rs
@@ -1,8 +1,8 @@
 fn main() {
     if true {
         let content = r#"
-fn main(arg: Foo<'static, 'static>) {
-}
+#[cfg(debug_assertions)]
+fn func() {}
 "#;
         let ast = syn::parse_file(content).unwrap();
         dbg!(ast);