+++ macro/src/util.rs
@@ -1,5 +1,6 @@
-use proc_macro2::Span;
 use std::path::PathBuf;
+
+use proc_macro2::{Span, TokenTree};
 use syn::{Error, ForeignItemFn, LitStr, Result};
 
 pub fn ident_from_pat(
@@ -11,14 +12,16 @@ pub fn ident_from_pat(
         syn::Pat::Ident(pat) => Ok(pat.ident.clone()),
         _ => Err(syn::Error::new(
             span,
-            format!("generating call for library function: signature of function {func_name} cannot be converted"),
+            format!(
+                "generating call for library function: signature of function {func_name} cannot be converted"
+            ),
         )),
     }
 }
 
 /// Reads the contents of a Rust source file and finds the top-level functions that have
 /// - visibility public
-/// - #[no_mangle] attribute
+/// - #[unsafe(no_mangle)] attribute
 ///   It converts these functions into a [syn::ForeignItemFn] so that those can
 ///   serve as lib function declarations of the lib reloader.
 pub fn read_functions_from_file(
@@ -29,7 +32,12 @@ pub fn read_functions_from_file(
     let path: PathBuf = file_name.value().into();
 
     if !path.exists() {
-        return Err(Error::new(span, format!("Could not find Rust source file {path:?}. Please make sure that you specify the file path from the project root directory. Please not that this has been changed in hot-lib-reloader v0.5 -> v0.6. See https://github.com/rksm/hot-lib-reloader-rs/issues/13.")));
+        return Err(Error::new(
+            span,
+            format!(
+                "Could not find Rust source file {path:?}. Please make sure that you specify the file path from the project root directory. Please not that this has been changed in hot-lib-reloader v0.5 -> v0.6. See https://github.com/rksm/hot-lib-reloader-rs/issues/13."
+            ),
+        ));
     }
 
     let content = std::fs::read_to_string(&path)
@@ -53,10 +61,25 @@ pub fn read_functions_from_file(
                     let no_mangle = fun
                         .attrs
                         .iter()
-                        .filter_map(|attr| attr.path.get_ident())
+                        .filter_map(|attr| attr.path().get_ident())
                         .any(|ident| *ident == "no_mangle");
 
-                    if !no_mangle {
+                    let unsafe_no_mangle = fun.attrs.iter().any(|attr| {
+                        attr.meta.require_list().is_ok_and(|list| {
+                            list.path
+                                .get_ident()
+                                .is_some_and(|ident| *ident == "unsafe")
+                                && list.tokens.clone().into_iter().any(|token| {
+                                    if let TokenTree::Ident(ident) = token {
+                                        ident == "no_mangle"
+                                    } else {
+                                        false
+                                    }
+                                })
+                        })
+                    });
+
+                    if !no_mangle && !unsafe_no_mangle {
                         continue;
                     };
                 }