+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 This package tries to adhere to [semver](https://semver.org/).
 
-## [0.8.0] (unreleased)
+## [0.8.0]
 ### Dependency updates, Rust 2024 edition, project cleanup
+This is an update that breaks backwards compatibility. hot-lib-reloader now user syn 2.0 and supports the Rust 2024 edition, in particular `#[unsafe(no_mangle)]` (https://github.com/rksm/hot-lib-reloader-rs/issues/42).
 
 ## [0.7.0]
 ### Add: Adjust the library filename
@@ -119,7 +120,7 @@ being used (b/c the prev solution wasn't really thread safe).
 
 ## [0.5.4]
 ### `#[no-mangle-if-debug]`
-Also add a [no-mangle-if-debug crate](https://github.com/rksm/hot-lib-reloader-rs/tree/master/macro-no-mangle-if-debug) that allows to `#[no_mangle]` functions but only in debug mode. The use of this is optional and nothing about hot-lib-reloader itself changes. This addresses https://github.com/rksm/hot-lib-reloader-rs/issues/10.
+Also add a [no-mangle-if-debug crate](https://github.com/rksm/hot-lib-reloader-rs/tree/master/macro-no-mangle-if-debug) that allows to `#[unsafe(no_mangle)]` functions but only in debug mode. The use of this is optional and nothing about hot-lib-reloader its<elf changes. This addresses https://github.com/rksm/hot-lib-reloader-rs/issues/10.
 
 
 ## [0.5.3]